CPython --> Cython, C++ --> C
Dealing with the CPython boilerplate code and Python 2 / 3 compatibility
was something I was not looking forward to. Cython made this much
easier.

Also, there were problems with compiling my alignment code (which used
C++) and the ksw2 code (which is pure C) together, so I dropped the
convenience of std::string so that everything could compile nicely
together in C.