New partition filtering strategy (#283)
The main motivation behind this PR was to discard unassemblable partitions *at the partitioning stage* to avoid wasting a lot of compute time trying to assemble many small poorly connected read sets. The intuition is that if it's a partition with a small number of reads and a small number of compatible overlaps, it can safely be discarded.

It turns out this intuition doesn't really pan out. There is no clear separation between assemblable and unassemblable partitions in terms of the number of edges in the read graph, the number of strict-mode edges in the read graph, or the ratio of the two.

So I've disabled the proposed filtering strategy, but left it commented out in the code for expository purposes. The refactoring in support of this filtering strategy was also valuable and has been retained.