Filter based on number of mismatches between contig and gDNA (#288)
In troubleshooting some false calls recently, I noticed that some are associated with contig alignments that have several mismatches with respect to the reference genome. A small number of off-target mismatches isn't unexpected, but if there are many off-target mismatches it severely reduces our confidence that the alignment is on-target and that the associated variant call(s) are correct.

This update deactivates the SNV detection procedure for any alignment with more than 4 mismatches between the contig and the reference. Preliminary testing has already shown a decrease in false positives without any decrease in sensitivity.