Refactor code for assembly, localization, and calling (#294)
Currently, kevlar optimizes the assembly, localization, and calling operations by composing several generator functions and invoking them with multiprocessing. A lot of the benefit is lost, however: the calls to BWA and fermi-lite do take advantage of multithreading, but the coordinating code is limited by GIL. Plus, there are a tremendous number of calls to BWA for localizing, which introduces a lot of overhead.

This update takes a different approach to optimizing these operations. The `kevlar assemble`, `kevlar cutout`, `kevlar call` commands now accept partitioned input files, and the `kevlar cutout` command (replacing `kevlar localize`) makes a single call to BWA rather than one call per partition. If multithreading is going to provide any benefit, it will be at the assembly stage, but that is out of scope for this update.