Logging improvements (#299)
A recent update introduced a new logging strategy that gives very frequent updates at first but then steadily gives less and less frequent updates. This helps give the user an idea of how quickly things are running at first, but in a way that doesn't subsequently overwhelm the user's terminal.

This update introduces a new `progress` module and `ProgressIndicator` class that isolates this behavior and makes it easy to use throughout the project. The new ProgressIndicator is now in use in the `assemble`, `call`, `cutout`, and `novel` modules.