Update pytest invocation and dependency version
This update makes sure that `kevlar/tests/__init__.py` is not included when invoking the test suite, as this causes problems (see https://github.com/pytest-dev/pytest/issues/4404).
With this update and commit dab6418b9f8979266dc3596d52d6131533e2fc2a, kevlar is now compatible with the latest version of pytest (4.0.0), and so the dependency version info has been updated in the `make devenv` target.

Closes #284.