Introduce --ctrl-max heurisitic at simlike stage (#327)
When searching for novel k-mers, the "--case-min" and "--ctrl-max" parameters are used to find k-mers that are putative signatures of novel variation: high abundance in the case sample (--case-min) and effectively absent (--ctrl-max) in the control samples. This update refines the workflow to filter out putative variants with a non-trivial number of spanning k-mers that do not satisfy the original selection criteria.

How many k-mers failing the selection criteria should be allowed before a putative variant is filtered out? In preliminary tests, there seems to be little effective difference between >2, >3, and >4. No negative effects have been observed, and the number of false calls is reduced dramatically.

The behavior is now as follows: the `ControlAbundance` filter is applied to any variant call with >4 "high-abundance" spanning k-mers in either parent/control sample. High-abundance is of course determined by the "--ctrl-max" parameter.

Fixes #323.