Speed up simlike by skipping calculations for filtered variants (#328)
By default, kevlar tries to provide as much information as possible with each variant call. For example, when scoring and ranking preliminary calls, kevlar will compute scores even for variants that have already been marked for filtering out. The hope is that this information might be of use if there is ever a need to troubleshoot results.

This update adds a new "fast mode" to the simlike module. In contrast to the default behavior, fast mode will avoid carrying out any computations it can for variants that have already been marked for filtering out. This provides a substantial speedup, which may be helpful when troubleshooting or during development.

Closes #324.