Filter based on proband k-mer abundance (#339)
It is often unreasonable to expect that all k-mers spanning the variant in the proband will be high abundance, and so kevlar has never required that all spanning k-mers are "interesting" or putatively novel. Even filtering based on the number of spanning k-mers that are "interesting" has been problematic, since some true variants sometimes have few spanning k-mers that meet the required thresholds.

However, kevlar has never filtered calls based on what is *not* expected in the k-mer spanning the variant in the proband. While it is common to have low abundance k-mers intermittently, it is very rare to have long stretches of low abundance k-mers spanning the variant in a true variant.

This update introduces a new filter that will discard a variant prediction if there are 5 or more low-abundance k-mers spanning the variant call. Based on observations, this will eliminate many false positives, while having little to no effect on true calls.

This update also makes the behavior of this new filter configurable, along with two other recently implemented filters.