Make --case|--controls|--refr args mandatory for simlike command
The `kevlar simlike` command supports an arbitrary number of VCF files as input. As a consequence, it's impossible to support the --case, --controls, and --refr arguments as positional arguments since --controls also supports an arbitrary number of values. The --case, --controls, and --refr settings are retained as option flags, but are now marked as mandatory. Closes #348.