Update VCF parser (#359)
This update modifies kevlar's VCF parser so that it does not choke on VCF produced by scalpel or DiscoSnp++. In particular, it checks whether VCF `FILTER` strings matched the pre-defined filters used by kevlar. If not, it prints a warning message rather than throwing an exception. This warning message can be disabled by setting `reader.suppress_filter_warnings` to `True`.

Closes #358.